requirements.txt
amqp==2.2.2
anyjson==0.3.3
apns2==0.3.0
appdirs==1.4.3
asn1crypto==0.22.0
Babel==2.3.4
bcrypt==3.1.3
billiard==3.5.0.3
boto==2.45.0
celery==4.1.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
coreapi==1.32.3
cryptography==2.0.3
defusedxml==0.5.0
Django==1.10.1
django-braces==1.11.0
django-celery==3.1.17
django-cors-headers==2.1.0
django-cors-middleware==1.3.1
django-filter==1.0.4
django-push-notifications==1.6.0
django-storages==1.5.2
django-suit==0.2.19
djangorestframework==3.6.3
djangorestframework-gis==0.12
drfdocs==0.0.11
easy-thumbnails==2.5
ecdsa==0.13
env==0.1.0
Fabric==1.13.2
Fabric3==1.13.1.post1
falcon==1.3.0
falcon-cors==1.1.7
fcm-django==0.2.16
gunicorn==19.6.0
h2==2.6.2
hpack==3.0.0
http-ece==1.0.5
hyper==0.7.0
hyperframe==3.2.0
idna==2.6
itypes==1.1.0
kombu==4.1.0
Markdown==2.6.8
numpy==1.14.0
oauthlib==2.0.2
olefile==0.44
openapi-codec==1.0.0
packaging==16.8
pandas==0.22.0
paramiko==2.2.1
Pillow==4.2.1
psycopg2==2.7.3
py-vapid==1.3.0
pyasn1==0.3.2
pycparser==2.18
pycrypto==2.6.1
pyfcm==1.4.5
PyJWT==1.5.2
PyNaCl==1.1.2
pyparsing==2.2.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
python-social-auth==0.2.21
python3-openid==3.0.10
pytz==2017.2
pywebpush==1.6.0
PyYAML==3.12
qrcode==5.3
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
simplejson==3.8.2
six==1.10.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4